// src/shared/enums/network/networkActionsType.enum.ts
export enum networkActionsType {
  confirmPendingRequest = 'confirm_pending_request',
  declinePendingRequest = 'decline_pending_request',
  canselSentRequest = 'cansel_sent_request',
  resendRequest = 'resend_request',
  resendFutureRequest = 'resend_future_request',
  canselFutureSentRequest = 'cansel_future_sent_request',
}